api/requirements.txt
Django==5.0.8
djangorestframework==3.15.2
psycopg[binary]==3.2.1
boto3==1.34.162
celery==5.4.0
redis==5.0.8
python-dotenv==1.0.1
numpy==2.0.1
scipy==1.14.0
matplotlib==3.9.1
